requirements.txt
# AISHA Project - Root Requirements
# Consolidating dependencies for AI services and system-wide python tools

# AI Services & Core dependencies
fastapi>=0.115.8
uvicorn==0.31.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
pydantic==2.9.2
pydantic-settings==2.5.2
requests>=2.32.4
sentence-transformers==3.1.1
reportlab==4.2.5
pypdf>=5.1.0
cryptography>=46.0.5
redis==5.1.1
apscheduler==3.11.0

# Development & Testing
pytest
pytest-asyncio
httpx

# External AI Integrations
ollama==0.4.7

# Utility Scripts (Root)
python-docx
python-multipart
beautifulsoup4
lxml
